Create register-login card component
